Fixed error Refactor Removed unused code
