+++ src/hdusd/utils/mx.py
@@ -19,7 +19,7 @@ import shutil
 
 from pathlib import Path
 
-from . import LIBS_DIR, title_str, code_str, BLENDER_VERSION
+from . import LIBS_DIR, title_str, code_str
 from .image import cache_image_file
 
 from . import logging
@@ -289,26 +289,3 @@ def export_mx_to_file(doc, filepath, *, mx_node_tree=None, is_export_deps=False,
 
     mx.writeToXmlFile(doc, filepath)
     log(f"Export MaterialX to {filepath}: completed successfuly")
-
-
-# update for material ui according to MaterialX nodetree header changes
-def update_material_iu(self, context):
-    if BLENDER_VERSION >= '3.0':
-        return
-
-    space = context.space_data
-    mat = context.object.active_material
-    if space.tree_type == 'ShaderNodeTree':
-        if space.node_tree != mat.node_tree:
-            space.node_tree = mat.node_tree
-            mat.hdusd.mx_node_tree = None
-            return
-
-    if space.tree_type != 'hdusd.MxNodeTree':
-        return
-
-    ui_mx_node_tree = mat.hdusd.mx_node_tree
-    editor_node_tree = space.node_tree
-
-    if editor_node_tree != ui_mx_node_tree and not space.pin:
-        mat.hdusd.mx_node_tree = editor_node_tree
